improve buildAll script

+++ buildAll.sh
@@ -1,8 +1,14 @@
 #!/usr/bin/env bash
 set -e
 localRepo=$(mvn help:evaluate -Dexpression=settings.localRepository -q -DforceStdout)
-rm -r "$localRepo/se/alipsa/matrix"
+if [[ -d "$localRepo/se/alipsa/matrix" ]]; then
+  echo "Removing local cache in $localRepo/se/alipsa/matrix"
+  rm -r "$localRepo/se/alipsa/matrix"
+fi
+echo "Building and locally publishing matrix"
 ./gradlew build publishToMavenLocal
+echo "Locally publish bom and verify"
 cd matrix-bom
 mvn -f bom.xml install
 mvn verify
+echo "Done!"
